Destructured the path plugin into 2 plugins
